Category select + Questions display
Sorted out the buttons for the category select. Now, on the selected category we can view the questions.

For now, going to do some refactoring to see what makes sense - do we keep everything in the questions view or do we bump it out to home view - and then working on the details view next